+++ project/Dependencies.scala
@@ -6,19 +6,19 @@ import sbt._
 object Dependencies {
 
   // Versions
-  private val SparkVersion = "2.3.1"
-  private val ExasolJdbcVersion = "6.0.8"
-  private val TypesafeLoggingVersion = "3.7.2"
+  private val SparkVersion = "2.4.0"
+  private val ExasolJdbcVersion = "6.0.13"
+  private val TypesafeLoggingVersion = "3.9.0"
 
   private val ScalaTestVersion = "3.0.5"
-  private val MockitoVersion = "2.23.0"
-  private val ContainersJdbcVersion = "1.10.1"
-  private val ContainersScalaVersion = "0.19.0"
+  private val MockitoVersion = "2.23.4"
+  private val ContainersJdbcVersion = "1.10.5"
+  private val ContainersScalaVersion = "0.22.0"
 
   private val sparkCurrentVersion =
     sys.props.get("spark.currentVersion").getOrElse(SparkVersion)
 
-  private val SparkTestingBaseVersion = s"${sparkCurrentVersion}_0.10.0"
+  private val SparkTestingBaseVersion = s"${sparkCurrentVersion}_0.11.0"
 
   val Resolvers: Seq[Resolver] = Seq(
     "Exasol Releases" at "https://maven.exasol.com/artifactory/exasol-releases"