+++ src/main/scala/com/exasol/spark/util/Types.scala
@@ -128,6 +128,54 @@ object Types extends LazyLogging {
     )
 
   /**
+   * Returns corresponding Jdbc [[java.sql.Types$]] type given Spark
+   * [[org.apache.spark.sql.types.DataType]] type
+   *
+   * @param dataType A Spark DataType (e.g. [[org.apache.spark.sql.types.StringType$]])
+   * @return A default JdbcType for this DataType
+   */
+  def jdbcTypeFromSparkDataType(dataType: DataType): Int = dataType match {
+    case IntegerType     => java.sql.Types.INTEGER
+    case LongType        => java.sql.Types.BIGINT
+    case DoubleType      => java.sql.Types.DOUBLE
+    case FloatType       => java.sql.Types.FLOAT
+    case ShortType       => java.sql.Types.SMALLINT
+    case ByteType        => java.sql.Types.TINYINT
+    case BooleanType     => java.sql.Types.BIT
+    case StringType      => java.sql.Types.CLOB
+    case BinaryType      => java.sql.Types.BLOB
+    case TimestampType   => java.sql.Types.TIMESTAMP
+    case DateType        => java.sql.Types.DATE
+    case dt: DecimalType => java.sql.Types.DECIMAL
+    case _               => throw new RuntimeException(s"Unsupported Spark data type $dataType!")
+  }
+
+  /**
+   * Returns corresponding Exasol type as a string for a given Spark
+   * [[org.apache.spark.sql.types.DataType]] type
+   *
+   * The types are obtained from Exasol manual, from a table named 'Summary of Exasol aliases',
+   * section 2.3.3 Data Type Aliases.
+   *
+   * @param dataType A Spark DataType (e.g. [[org.apache.spark.sql.types.StringType$]])
+   * @return A default Exasol type as string for this DataType
+   */
+  def exasolTypeFromSparkDataType(dataType: DataType): String = dataType match {
+    case ShortType       => "SMALLINT"
+    case ByteType        => "TINYINT"
+    case IntegerType     => "INTEGER"
+    case LongType        => "BIGINT"
+    case DoubleType      => "DOUBLE"
+    case FloatType       => "FLOAT"
+    case dt: DecimalType => "DECIMAL"
+    case BooleanType     => "BOOLEAN"
+    case StringType      => "CLOB"
+    case DateType        => "DATE"
+    case TimestampType   => "TIMESTAMP"
+    case _               => throw new RuntimeException(s"Unsupported Spark data type $dataType!")
+  }
+
+  /**
    * Select only required columns from Spark SQL schema
    *
    * Adapted from Spark JDBCRDD private function `pruneSchema`.
@@ -143,6 +191,25 @@ object Types extends LazyLogging {
     logger.debug(s"Using a new pruned schema $newSchema")
     newSchema
   }
+
+  /**
+   * Returns comma separated column name and column types for Exasol table from Spark schema
+   *
+   * @param schema A Spark [[org.apache.spark.sql.types.StructType]] schema
+   * @return A comma separated column names and their types
+   */
+  def createTableSchema(schema: StructType): String =
+    schema.fields
+      .map { field =>
+        val fieldType = Types.exasolTypeFromSparkDataType(field.dataType)
+        val nameType = s"${field.name} $fieldType"
+        if (field.nullable) {
+          nameType
+        } else {
+          nameType + " NOT NULL"
+        }
+      }
+      .mkString(", ")
 
   def getMaxPrecisionExasol(): Int =
     MAX_PRECISION_EXASOL