+++ src/test/scala/com/exasol/spark/ExasolRelationSuite.scala
@@ -2,7 +2,6 @@ package com.exasol.spark
 
 import org.apache.spark.rdd.RDD
 import org.apache.spark.sql.Row
-import org.apache.spark.sql.SQLContext
 import org.apache.spark.sql.sources._
 import org.apache.spark.sql.types._